+++ tools/apicheck/stubs/io/papermc/paper/registry/RegistryAccess.java
@@ -1,0 +1,9 @@
+package io.papermc.paper.registry;
+
+import org.bukkit.Keyed;
+import org.bukkit.Registry;
+
+public interface RegistryAccess {
+    static RegistryAccess registryAccess() { throw new UnsupportedOperationException(); }
+    <T extends Keyed> Registry<T> getRegistry(RegistryKey<T> registryKey);
+}